requirements.txt
numpy
sounddevice
soundfile
scipy
pyinstaller
tokenizers
onnxruntime-directml
# Discord bot dependencies
discord.py[voice]
davey
